requirements.txt
langchain 
wikipedia 
langchain-community 
langchain-openai
langchain-anthropic
langchain-google-genai
python-dotenv
pydantic
duckduckgo-search
fastapi
uvicorn
python-multipart
python-jose[cryptography]
firebase-admin
google-generativeai
google-genai